// src/app/models/InformeTecnico.ts
export interface InformeTecnico {
  idinformes_tecnicos: number;
  informes_tecnicos_fecha_recibido: string;
  informes_tecnicos_tipo_compra: string;
  informes_tecnicos_nro: string;
  informes_tecnicos_todo: string;
  informes_tecnicos_detalles: string;
  informes_tecnicos_estado: string;
  informes_tecnicos_tecnico: string;
}